+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/Maintenance/EditTradingAccountCommand.cs
@@ -2,7 +2,7 @@ using System.Windows;
 using TD.WPF.Services;
 using TD.WPF.ViewModels;
 
-namespace TD.WPF.Commands;
+namespace TD.WPF.Commands.Maintenance;
 
 internal class EditTradingAccountCommand : AbstractCommand
 {