+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Services/WindowDialogService.cs
@@ -8,6 +8,18 @@ namespace TD.WPF.Services;
 internal class WindowDialogService : IWindowDialogService
 {
     /// <inheritdoc />
+    public (bool IsConfirmed, Currency? Currency) ShowCurrencyCreateDialog(Window owner)
+    {
+        var dialog = new CurrencyCreate()
+        {
+            Owner = owner
+        };
+
+        var result = dialog.ShowDialog() == true;
+        return (result, result ? dialog.Currency : null);
+    }
+
+    /// <inheritdoc />
     public (bool IsConfirmed, bool IsDirty) ShowCurrencyEditDialog(Window owner, Currency currency)
     {
         var dialog = new CurrencyEdit(currency)
@@ -20,6 +32,18 @@ internal class WindowDialogService : IWindowDialogService
     }
 
     /// <inheritdoc />
+    public (bool IsConfirmed, Exchange? Exchange) ShowExchangeCreateDialog(Window owner)
+    {
+        var dialog = new ExchangeCreate()
+        {
+            Owner = owner
+        };
+
+        var result = dialog.ShowDialog() == true;
+        return (result, result ? dialog.Exchange : null);
+    }
+
+    /// <inheritdoc />
     public (bool IsConfirmed, bool IsDirty) ShowExchangeEditDialog(Window owner, Exchange exchange)
     {
         var dialog = new ExchangeEdit(exchange)
@@ -29,6 +53,42 @@ internal class WindowDialogService : IWindowDialogService
 
         var result = dialog.ShowDialog() == true;
         return (result, dialog.IsDirty);
+    }
+
+    /// <inheritdoc />
+    public (bool IsConfirmed, Symbol? Symbol) ShowSymbolCreateDialog(Window owner, IExchangeLookupService exchangeLookupService, ICurrencyLookupService currencyLookupService)
+    {
+        var dialog = new SymbolCreate(exchangeLookupService, currencyLookupService)
+        {
+            Owner = owner
+        };
+
+        var result = dialog.ShowDialog() == true;
+        return (result, result ? dialog.Symbol : null);
+    }
+
+    /// <inheritdoc />
+    public (bool IsConfirmed, bool IsDirty) ShowSymbolEditDialog(Window owner, Symbol symbol)
+    {
+        var dialog = new SymbolEdit(symbol)
+        {
+            Owner = owner
+        };
+
+        var result = dialog.ShowDialog() == true;
+        return (result, dialog.IsDirty);
+    }
+
+    /// <inheritdoc />
+    public (bool IsConfirmed, TradingAccount? TradingAccount) ShowTradingAccountCreateDialog(Window owner)
+    {
+        var dialog = new TradingAccountCreate()
+        {
+            Owner = owner
+        };
+
+        var result = dialog.ShowDialog() == true;
+        return (result, result ? dialog.TradingAccount : null);
     }
 
     /// <inheritdoc />
@@ -47,18 +107,6 @@ internal class WindowDialogService : IWindowDialogService
     public (bool IsConfirmed, bool IsDirty) ShowTradingAccountEditDialog(Window owner, TradingAccount tradingAccount)
     {
         var dialog = new TradingAccountEdit(tradingAccount)
-        {
-            Owner = owner
-        };
-
-        var result = dialog.ShowDialog() == true;
-        return (result, dialog.IsDirty);
-    }
-
-    /// <inheritdoc />
-    public (bool IsConfirmed, bool IsDirty) ShowSymbolEditDialog(Window owner, Symbol symbol)
-    {
-        var dialog = new SymbolEdit(symbol)
         {
             Owner = owner
         };